fix double

+++ src/com/company/Addition.java
@@ -1,9 +1,9 @@
 package com.company;
 
 public class Addition extends Operation{
-    int add;
+    double add;
 
-    public Addition(int number1, int number2){
+    public Addition(double number1, double number2){
 
         super( number1, number2, '+');
         this.add = number1 + number2;
