+++ src/com/company/Multiplication.java
@@ -1,9 +1,9 @@
 package com.company;
 
 public class Multiplication extends  Operation{
-    int mul;
+    double mul;
 
-    public Multiplication(int number1, int number2){
+    public Multiplication(double number1, double number2){
 
         super( number1, number2, '*');
         this.mul = number1 * number2;
